Create Team Fresh 7 Segment Display (Draft)
This is more or less a template for the Seven Segment display. This code is compatible with a certain hardware (complicated name).This code won't work with just the seven segment display itself with an arduino or other logic board. A wiring diagram is available at this site: scuola.arduino.cc